feature - cart: Crud feature

+++ src/main/java/com/example/cecv_e_commerce/domain/dto/cart/CartDTO.java
@@ -2,5 +2,5 @@ package com.example.cecv_e_commerce.domain.dto.cart;
 
 import java.util.List;
 
-public record CartDTO(Long id, Long userId, List<CartItemDTO> items) {
+public record CartDTO(Integer id, Integer userId, List<CartItemDTO> items) {
 }
