+++ src/main/java/com/example/cecv_e_commerce/repository/CartRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.cecv_e_commerce.domain.model.Cart;
 
-public interface CartRepository extends JpaRepository<Cart, Long> {
-    Optional<Cart> findByUserId(Long userId);
+public interface CartRepository extends JpaRepository<Cart, Integer> {
+    Optional<Cart> findByUserId(Integer userId);
 }
